Add files via upload
This is the .py file and you can launch it in the way you would launch a normal python program. You can use cmd or command line to do it or depending on what python you have setup you can just open it. You can also make changes to the code as you'd like.